restrictd more delete events fltering

+++ src/api/routes/kubernetes.js
@@ -11,11 +11,14 @@ module.exports = {
   'kubernetes#event': async (req, res, next) => {
     logger.log('received kubernetes event');
     try {
-      const { reason, message, involvedObject: { name, namespace } } = req.body;
-      logger.log(`[${reason}] received kubernetes event: ${message} ${name} in ${namespace}`);
+      const {
+        reason, message, type, involvedObject: { name, namespace },
+      } = req.body;
+      logger.log(`[${reason}] received ${type} kubernetes event: ${message} ${name} in ${namespace}`);
 
       // remove only pods in your namespace and due to backoff errors
-      if ((namespace.match('^pipeline-.*') || namespace.match('^worker-.*')) && reason === 'BackOff') {
+      if ((namespace.match('^pipeline-.*') || namespace.match('^worker-.*'))
+       && reason === 'BackOff' && type !== 'Normal' && message.contains('restarting')) {
         const k8sApi = kc.makeApiClient(k8s.CoreV1Api);
         logger.log(`removing pod ${name} in ${namespace}`);
         await k8sApi.deleteNamespacedPod(name, namespace);
